add wait timeout to sync AxiDma
This adds wait_d2h_timeout and wait_h2d_timeout methods to AxiDma.
These methods wait for an interrupt to be received until a timeout
is reached. If the timeout is reached without an interrupt being
received, then the method returns Error::Timeout. If an interrupt
is received, then the method returns Ok(()).

The crate polling is used as a portable way to use epoll on the
UIO file descriptor to check if it is ready for reading (however
note that UIO is Linux only, so portability is not an important
need here).

All these changes are gated behind a feature called "timeout", so
that code that does not need to use these timeouts does not need
to bring in polling as a dependency.